+++ src/Domain/Pattern/Services/RepgeneratorStaticFilesService.php
@@ -49,11 +49,12 @@ class RepgeneratorStaticFilesService
             $name = explode('.', $nameWithExtension)[0];
             $file = $this->getStatic($name);
 
-            if (!file_exists($path = app_path($fileOriginal))) {
+            //if (!file_exists($path = app_path($fileOriginal))) {
+                $path = app_path($fileOriginal);
                 file_put_contents($path, $file);
 
                 $generatedFiles[] = new RepgeneratorStaticFileAdapter($nameWithExtension, $path);
-            }
+            //}
         }
         return $generatedFiles;
     }